Add files via upload
This is the original upload of files that may change, this version just works on creating a message that ask for a name. Additionally, dynamic programming was implemented for the first time, image is atomatic(the same one). And this is just to have somewhere to fall back if all else fails